+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
+## [0.1.2]
+
+### Fixed 
+- identity-box not using system hash code
+
 ## [0.1.1]
 
 ### Fixed